requirements.txt
flask
google-generativeai
flask-cors
chromadb
langchain
sentence-transformers
python-dotenv
